Optimize execution of OR in AIJ and ATJ.
Now AIJ supports the execution of a "primitive pseudo-function" OR that
evaluates its arguments non-strictly. This does not interfere with the ACL2 OR
macro because AIJ deals with translated terms, where macros are already
expanded, and furthermore ACL2 prohibits the definition of functions whose names
are, like OR, in the "COMMON-LISP" package.

Now ATJ recognizes (translated) terms of the form (IF A A B), likely to come
from untranslated terms of the form (OR A B), and generates code to construct
AIJ representations of (OR A B) (where here OR is the primitive pseudo-function
mentioned above).

Thanks to Keshav Kini for bringing up the issue at the ACL2-2018 Workshop.