Start formalizing Ethereum's cryptographic interface.
This is similar to the cryptographic interface in the Bitcoin library, namely a
collection of weakly contrained ACL2 functions that the Ethereum formalization
is parameterized over. Actual definitions of these cryptographic functions can
be DEFATTACHed to these weakly contrained functions.

We start with Keccak-256. Others will be added as needed.