Reduce sizes of make-event expansions in .cert files.
Quoting :doc note-8-2:

  [Make-event] expansions have often been reduced in size.  (Technical
  note: record-expansion has been inserted only on expansions done
  directly under [encapsulate] events, and some changes have also
  been made in how and when [local] events are elided from
  expansions.)  The reduction in total sizes of .cert files for a
  complete (``everything'') regression was 8.8%, though in some cases
  the reduction was substantially larger: for example, the size of
  books/centaur/fty/tests/deftranssum.cert was reduced from
  22,474,113 bytes to 16,910,530 bytes, a reduction of nearly 25%.

Perhaps the main benefit of this work is simplification and
clarification of the use of record-expansion and elide-locals in the
implementation of make-event.  See the Essay on Make-event (which is
in the #+acl2-loop-only definition of make-event in axioms.lisp).

Normally, any size reduction in .cert files will only be modest.
However, if the input form is large, in particular involving large
constants using #. syntax, the reduction can be substantial.  Consider
for example a book with portcullis command

  (defconst *c* (make-list 1000))

and event:

  (make-event `(defun foo () '#.*c*))

The size in bytes of foo.cert was reduced from 18877 to 3698.