Added a blank line to help the build system past an obstacle.
Added a blank line to cause arithmetic-5/lib/basic-ops/top.lisp to be
recertified, as well as arithmetic-5/top.lisp (which depends on it).
A better solution is to use the capability documented in :doc
build::acl2-system-feature-dependencies, but I need to give that a bit
of thought so for now, I'm taking the easy way out.