Updated ccl installation instructions.  Installed bug fix for guard verification involving lambdas.
We failed to check that the lambda body was composed only of guard
verified functions.  This fix is from J, as is the following example,
which formerly caused the raw Lisp error shown at the end.

(include-book "projects/apply/apply-lemmas" :dir :system)

(defun$ sq (n)  ; Note the unverifiable guard!
  (declare (xargs :guard (or (eq n 'nan) (natp n))
                   :verify-guards nil))
  (* n n))

(defun test ()
  (declare (xargs :guard t))
  (apply$ (lambda$ (k)
                   (declare (xargs :guard (or (eq k 'nan) (natp k))))
                   (sq k))
          '(nan)))

(test)

***********************************************
************ ABORTING from raw Lisp ***********
********** (see :DOC raw-lisp-error) **********
Error:  The value NAN is not of the expected type NUMBER.
While executing: CCL::*-2
***********************************************