[X86ISA] Improve checks in segment address translation.
The function EA-TO-LA, which translates an effective address (given a segment
register) into a linear address, was checking that the effective address is
within segment limits (in 32-bit mode), but then functions like RME-SIZE were
accessing potentially other effective addresses past that one.

Now EA-TO-LA takes an extra argument for the number of bytes that will be
accessed, starting from the effective address. The revised EA-TO-LA checks that
the whole chunk, from EFF-ADDR to EFF-ADDR + NBYTES - 1, is contained in the
segment.

There is still a complication that needs to be addressed, namely that the
addition EFF-ADDR + (NBYTE - 1) should be modular presumably, which may result
in two separate chunks of addresses that have to be checked for containment into
the segment. But operations like RME-SIZE always read one chunk, by calling
RML-SIZE on the translated (first) address, so RME-SIZE and similar operations
may need to be revised as well.

The current check performed by EA-TO-LA should be "safe" in the sense that it is
more strict: it rejects cases where EFF-ADDR + (NBYTES + 1) wraps around (which
should be a rare/strange occurrence anyhow).