Started new :doc efficiency topic.
Quoting :doc note-8-2:

  A new documentation topic, [efficiency], suggests some ways to speed
  up proofs and evaluation.  The ACL2 community is encouraged to
  extend (and more generally, improve) this topic!

NOTE: It will be great if others contribute.  I think I'll announce
the existence of this topic on the ACL2 mailing list around Jan. 7.
If you want me to wait beyond that because you plan to contribute,
just ask.

In particular, I haven't added links to :doc efficiency in the manual,
except in :doc note-8-2.  Feel free to do so!

Thanks to Alessandro and Eric S. for their initial feedback.

In particular, Alessandro brought up the question of whether it's
inefficient to hang many rules on the same function symbol.  I seem to
recall that Sol ran an experiment that suggested it's not a big deal,
so I searched old emails and the ACL2 sources, but didn't find
anything about that.  So I haven't addressed that topic in :doc
efficiency.

By the way, I almost added the following at the end of the "Proof
efficiency" section.  But I don't know if there's much truth to it.
Maybe someone else has an example?

 <p>For extremely long proofs, it might be worthwhile to turn off prover
 output, e.g., @('(set-inhibit-output-lst '(proof-tree prove
 proof-builder))').  Since @(see gag-mode) is on by default, this probably
 won't be important for reducing prover output; however, gag-mode stores proof
 output for later replay using utilities like @(':')@(tsee pso), so turning off
 prover output can save a lot of memory, which could save some time.</p>