[X86ISA] Expand some documentation.
Based on discussions with Shilpi Goel, and on the article she found on the web,
I've revised the part of the documentation of EA-TO-LA about the issue of
whether the chunk of effective addresses can wrap around or not. The article
states that it cannot, but we agreed that we still need to run some experiments.
So I've revised the documentation to acknowledge the still somewhat open issue
and to reference the article.

Note that the current check for containment in the segment is "safe" in the
sense that it is possibly more strict than it needs to be, but the difference
should be observable only in very pathological cases.