[X86ISA] Fix a bug in SYSCALL and SYSRET.
These instructions were setting the G (granularity) bit of the code or stack
segment to 1, but the limit to just #xFFFFF (20 bits) instead of #xFFFFFFFF (32
bits). As noted in utils/segmentation-structures.lisp, in the model's
representation of the hidden part of segment registers, the limit field must be
already "processed" according to the G bit: that is, left unchanged if G = 0,
and scaled by 12 bits if G = 1.