[ALISTS] Add ALIST-MAP-KEYS and ALIST-MAP-VALS.
These two functions are similar to STRIP-CAR and STRIP-CDRS, as well as to
ALIST-KEYS and ALIST-VALS, but they ignore shadowed pairs. That is, they operate
on the mathematical map represented by the alist. Crucially, ALIST-MAP-VALS
returns the range or image (depending on terminology) of the map, which neither
STRIP-CDRS nor ALIST-VALS do precisely if there are shadowed pairs.
ALIST-MAP-KEYS returns something equivalent to STRIP-CARS and ALIST-KEYS, but it
is a natural companion to ALIST-MAP-VALS.

Thanks to Matt Kaufmann, Keshav Kini, and Grant Passmore for useful Slack
discussions about how to name these functions. The chosen names were proposed by
Grant Passmore.