[X86ISA] Handle the CS.R bit.
The R bit of a code segment descriptor says whether the code segment is
execute-only (R = 0) or is also readable (R = 1). In 32-bit mode, when reading
data (as opposed to code, including immediate operands) from the code segment,
this bit must be 1, otherwise access is not allowed.

So the R(I)MExx and R(I)ME-SIZE functions have been extended to include this
check.