[X86ISA] Fix bug in SEGMENT-BASE-AND-BOUNDS.
This function was applying data segment field selectors to all segments, not
just data segments. It has been now changed to use code segment field selectors
for code segments. While the D and D/B bits of code and data segments are in the
same position, reading the E bit from a code segment was incorrect, since a code
segment has no E bit (the C bit is in the same position).

Thanks to Eric Smith for discussion that led to the discovery of this issue.