[AIJ] Optimize execution via variable indices.
The Java representation of ACL2 variables has been optimized to include numeric
indices derived from the position of variables in the lists of parameters in
named function definitions and lambda expressions. These indices are set in the
body of each function definition as the function definition is added to the Java
representation of the ACL2 environment.

The AIJ evaluator has been optimized to use Java arrays of ACL2 values as
bindings, instead of hash maps from ACL2 symbols to values, enabling faster
lookup of variables in bindings.

These optimizations provide roughly a 2x speedup on the examples in the
ACL2-2018 Workshop paper (except for the factorial example, where big integer
multiplication dominates).

These optimizations are internal to AIJ (the embedding). No change was required
for ATJ (the code generator).