[GL & X86ISA] Move and rename GL utility in X86ISA.
The DEF-GL-EXPORT in the x86 ISA model is more general than the x86. Thus, it
has been moved into a new file under GL, along with XDOCumentation. It has been
renamed to DEFTHM-USING-GL, as suggested by Shilpi Goel. The use of MK-NAME (a
utility in the x86 model) has been replaced with the built-in ADD-SUFFIX system
utility.