Made another fix to ev-fncall-rec-logical, for attachments, together with raw-ev-fncall-okp.  Improved some code to use record accessors for apply$-badge.  Tweaked some comments about badges and warrants.
Fixed second buggy fix to source function ev-fncall-rec-logical.
Accordingly tweaked raw-ev-fncall-okp in support of checking done for
ev-fncall-rec.  So now, quoting a comment in (defxdoc note-8-2 ..):

; Fixed ev-fncall-rec-logical and raw-ev-fncall-okp to deal properly with
; attachments to warrants.

Tweaked theorem concrete-badge-userfn-type and the definitions of
apply$-badgep and concrete-apply$-userfn to use record accessors for
apply$-badge.

Fixed some comments about badges and warrants in consultation with J.

Note: I did only a partial test of this commit as shown below (as well
as "make proofs"), so I am not pushing this to GitHub yet.

./books/build/cert.pl -j 8 --acl2 <my_acl2> books/system/apply/*.lisp books/projects/apply/*.lisp books/system/tests/loop-tests.lisp