[SOFT] Removed explicit function parameters.
The DEFUN2, DEFCHOOSE2, DEFUN-SK2, and DEFUN-INST no longer include an explicit
list of function parameters. The function is implicitly parameterized over the
function variables that it depends on.

This simplified the implementation and avoids the repetition of function
variables when using the naming convention of putting the function variables,
enclosed in square brackets, in the name of the function.