[BITCOIN] Formalize the BIP 32 wallet structure.
The recommended wallet structure in BIP 32 is formalized via a predicate that
constrains the shape of the tree. The definition is complicated by the rare but
mathematically possible cases in which certain keys are invalid and need to be
skipped.

With this, BIP 32 is now completely formalized. (It can be improved, of course,
like everything else.)