Add XDOC topic(s) that show the built-in theorems.
Here 'theorem' means both DEFAXIOMs and DEFTHMs (i.e. theorems in the logical
sense). These topics contain sequences of @(def ...) references to all the
built-in theorems (from axioms.lisp), except those that involve internal
(i.e. non-documented) functions. It may be easier to find and examine the
built-in theorems in these XDOC topics than in axioms.lisp, which contains a lot
of additional code.

The division into subtopics is preliminary. We could have finer-grained and
perhaps hierarchical subtopics.

This, or an evolution of this, might be incorporated into the ACL2 manual
itself, if there's consensus.