[AIJ & ATJ] Improve implementation.
Now all the defined functions are stored into a static field of the
Acl2NameFunction class, instead of in the Acl2Environment class. The methods to
define and call named functions has been accordingly moved to the
Acl2NamedFunction class from Acl2Environment.

Adjust ATJ to generate slightly different code, to accommodate the change to AIJ
described in the previous paragraph.