[X86ISA] Improve PUSH segment register.
When the operand size is 32 bits, now the model of this instruction writes the
low 16 bits, leaving the high 16 bits unchanged. This is what modern processors
do, while older ones zero-extend the 16-bit value to 32 bits and then write all
32 bits to the stack.

In the future, we may parameterize the model over a flag that determines the
behavior of PUSH segment register with a 32-bit operand size -- modern or
legacy.