[XDOC] Fix bug in some XDOC constructors.
This has to do with the constructors TABLE, U, and SEE, which would conflict
with existing functions and thus have to be named slightly differently. However,
the tag keywords should still be called :TABLE, :U, and :SEE, so this is now the
case.

I discovered this bug while checking the examples in my recent email to ACL2
Books, so thanks to David Rager, Eric McCarthy, and Eric Smith for the
discussion, also for leading to the discovery of this bug.