Improved an apply$ rule to avoid harmful forcing.  Misc. :doc improvements.
I discovered a problem with forcing done by rule ev$-opener
books/projects/apply/base.lisp.  Thanks to J for thinking through the
(as far as we know) best solution and explaining it with a comment
just below the changed rule; these are now in the aforementioned
base.lisp book.

Misc. :doc improvements:
- Updated :doc developers-guide-contributing:
  - Optionally also send the commit hash for the starting point
    (thanks to Mihir Mehta for noticing that this was missing).
  - Final step includes "git pull".
- Added description of make-lambda-application in :doc system-utilities,
  with comparison to make-lambda-term.
- Tweaked the formatting of the General Form in :doc loop$ to look a
  bit better than before at the terminal and in acl2-doc, and also
  fixed a typo (``a a'') there.