[Digits-Any-Base] Add a macro DEFDIGITS.
This macro specializes the operations in the Digits-Any-Base library to a
specified base and to specified unary recognizers and fixers for digits. Thus, a
user can take existing types whose values are digits in a particular base
(e.g. bytes, which are digits in base 256), and generate specialized operations
to convert between big-endian and little-endian lists of bytes and natural
numbers. Theorems are also generated, by specializing (some of) the general ones
in the library. Overall, this should support more automated and streamlined
proofs involving these operations.