Fixed thm to respect skip-proofs environments.  Made rewrite-rule-term executable.  Added a comment in pkg-names-memoize.
Thm was doing a proof even when proofs were skipped, except (probably)
for when including a certified book.  That has been fixed.

Rewrite-rule-term is now a guard-t function that wraps ec-call around
a call of a new function, rewrite-rule-term-exec, which is like the
old rewrite-rule-term but is executable with a proper guard.  Quoting
a comment in :doc note-8-2 (I wonder if anyone will notice this):

; The function rewrite-rule-term-exec is an executable version of
; rewrite-rule-term, which is now executable.

Added a comment in pkg-names-memoize about a stack overflow issue.