Reverted attempt to improve ACL2(p) behavior using ccl::with-terminal-input.
I've reverted the use of ccl::with-terminal-input, as explained in
this revised comment in (defxdoc note-8-2 ...).  With that change, an
ACL2(p) regression passed without error (though a previous run had one
hang, so I don't think ACL2(p) is 100% reliable yet), using a version
of CCL with a bug fix (see https://github.com/Clozure/ccl/issues/207)
that isn't yet in CCL master.  Note that the hang mentioned below
happens not only with that version of CCL but also with one that I
built in February.

; ACL2(p) built on CCL could occasionally hang during proofs because of
; contention by two threads when reading terminal input.  We attempted to make
; this less likely by using ccl::with-terminal-input.  We thank Michał Herda
; for diagnosing this problem and David Rager for helpful discussions.
; Unfortunately, this change caused a repeatable hang in community book
; books/std/osets/under-set-equiv.lisp (and there may be problems for other
; books).  We have left the macro our-with-terminal-input in place, documenting
; our current (lack of) understanding, but it is the identity macro at this
; point.