[Digits-Any-Base] Add ops to group and ungroup digits.
These operations group digits in a smaller base into digits in a larger base,
and vice versa. There are both big-endian and little-endian variants.

These can be used, for instance, to convert between nibbles and bytes, or
between bits and bytes, or between numbers below 10 and numbers below 100,
etc. It allows conversions between any base and any positive power of it.