Improved :doc loop$ based on discussions with J.  Added generated files doc/acl2-code-size.txt and doc/home-page.html to the GitHub repository.
While it is perhaps unusual to add generated files to a GitHub
repository, I think it makes sense in this case.  I already use "make
DOC" to keep generated file doc.lisp up-to-date, and that should keep
these other two files up-to-date as well.  I ran this by the
acl2-books list and got some assent and no objection; J was fine with
this, too.