Dummy commit to acknowledge addition of doc/acl2-code-size.txt and doc/home-page.html to the repository.
[I was concerned that --allow-empty would do something surprising, so
I just added a blank line to the end of doc.lisp to do a dummy
commit.]

As discussed on the acl2-books list, these two generated files are
suitable for the repository.  Note that we already have a generated
file in the repository for good reason, doc.lisp, which has been there
for a long time.  So maintaining those two files would be easy.
What's more, doc/home-page.html would probably change only rarely.

Thanks to Warren Hunt for noticing the absence of doc/home-page.html
from the GitHub release tarball.  This will fix that problem.