[FTY] Improve DEFBYTE.
If a predicate with the name of the recognizer already exists, use that without
re-generating it. See the updated documentation and implementation for details.

In particular, this avoids the shadowing of the XDOC topics BYTEP and NIBBLEP
(in [books]/std/basic/bytep.lisp and [books]/std/basic/nibblep.lisp) by the
calls of DEFBYTE in [books]/kestrel/fty/byte.lisp and
[books]/kestrel/fty/nibble.lisp. Thanks to Matt Kaufmann for finding that these
topics were being shadowed.