[Std/lists] Remove TAKE-REDEFINITION and TAKE-INDUCTION.
Now that the built-in definition of TAKE has been changed to be essentially like
TAKE-REDEFINITION in Std, there is no need for this rule to remain, which in
fact was slowing down some proofs. The companion rule TAKE-INDUCTION, an
induction rule based on TAKE-REDEFINITION, has also been removed, because it
should be now equivalent to the built-in induction rule generated by the new
definition of TAKE.

Several books had to be adapted to this change. Occurrences of TAKE-REDEFINITION
(e.g. in hints) have been replaced by TAKE. It was also necessary to enable TAKE
in several proofs that were otherwise failing. An :EXPAND hint also had to be
added to one proof, to force expansion of TAKE in the inductive case of the
proof. In a few places, it was necessary to enable or disable other TAKE-related
rules, such as TAKE-WHEN-ATOM.

While the changes to the books were minimal and all the proofs work, book
authors may want to double-check their files and perhaps tweak things.