Adapt community books to changes to DEFSTUB.
DEFSTUB, in the system code, has been extended to accept the same keywords as
ENCAPSULATE, for both new-style and old-style signatures. This commit adapts the
community books to that extension.

:DOC DEFSTUB has been modified to reflect the extension.

:DOC NOTE-8-3 has been extended to mention the extension to DEFSTUB.

The relevant to-do item in the developers' to-do list has been removed. That
item concerned just the addition of :GUARD to DEFSTUB, but while at it, it makes
sense to support everything that ENCAPSULATE supports, so that DEFSTUB is just
an "optimized" version of ENCAPSULATE when there are no constraints (except for
the TRUE-LISTP property of functions returning MV).

Tests for DEFSTUB have been added. Strictly speaking this is not a necessary
adaptation to the community books, but it certainly goes well with the extension
to DEFSTUB.