[X86ISA] Have DEF-INST bind CTX.
The DEF-INST macro has been extended to automatically generate a binding of the
function name to the CTX variable. This makes each uses of DEF-INST a little
shorter and prevents the function name and the value of CTX from getting out of
sync.

This is a simple first step towards having DEF-INST generate more of the
boilerplate code in the instruction semantic functions.