[DEFARBREC] Add a new macro, DEFARBREC.
This macro (which stands for 'define arbitrary recursion') enables the
introduction of recursive functions without having to prove termination right
away. There exist similar tools in the community books, but the specifics of
this macro are motivated by its intended use with APT. See :DOC
DEFARBREC-RELATED for more information on how this macro relates to the existing
ones.