[APT] Extend RESTRICT to support reflexive functions.
These are recursive functions that occur in their termination theorem.

The generation of the termination proof for the new function takes advantage of
the automatic functional instantiation of (:TERMINATION-THEOREM ...) that was
implemented to support this use case.

In order to generate a proof for the correctness of the new function, first the
key applicability condition is slightly strengthened by replacing occurrences of
the old function with a freshly and locally generated stub with the same arity
(a no-op if the function is not reflexive), and then the applicability condition
is functionally instantiated (in the correctness proof for the new function) by
replacing the stub with the new function (again, a no-op if the function is not
reflexive).