Fix books for the addition to meta-extract-global-facts of a check for logicp of the function called by magic-ev-fncall.
When using magic-ev-fncall in a verified clause processor or meta rule
and using meta-extract to assume that the result is correct, we now
need to know logicp of the function symbol passed to magic-ev-fncall.
A fix is to replace such calls
 (magic-ev-fncall fn args state t nil)
with
 (magic-ev-fncall-logic fn args state).
This is just a wrapper that performs the logicp check itself.