[Java] Fix a bug in ATJ.
This is in the new shallow embedding code generation part.

When recognizing IFs of the form (IF a a b) to treat them as (OR a b) during
execution, the code was passing the second a, instead of b, as the second OR
argument to the OR code generation function.