[Java] Use Java abstract syntax and pretty-printer in ATJ.
Now ATJ includes a more general abstract syntax for Java and associated
pretty-printer, which it uses for code generation, instead of writing directly
to the output channel piecewise as before.

The abstract syntax covers a large part of Java, certainly all the constructs
targeted by ATJ. The pretty printer generates lines as MSGP values, using
recursively formatted strings which make printing code more readable. There is
just one function that writes to the output channel, writing all the lines
produced by the rest of the pretty-printer.

This way, all of the abstract syntax, most of the pretty-printer, and more of
the rest of ATJ can be in logic mode and guard-verified (although this has not
been done for all the relevant parts of ATJ yet), providing type-checking-like
guarantees -- the abstract syntax is built with FTY.