[Java] Add a :GUARDS option to ATJ.
This option, when T, generates Java code that assumes that the guards are always
satisfied. Currently the only difference that this option makes is that Java
code is generated only for the :EXEC parts of MBE, instead of only for the
:LOGIC parts, but it opens the door to faster execution under the guards.

This new :GUARDS option should be used only with guard-verified ACL2 code and
with external Java code that calls the generated Java code always with values
that satisfy the guards.

This is just a start towards handling guards in ATJ (and AIJ); more
sophisticated support will be added in the future.