[Java] Improve AIJ.
Narrow the return type of some of the methods that provide native Java
implementations of the ACL2 primitive functions, e.g. the method for EQUAL now
returns Acl2Symbol instead of Acl2Value.

This is also in preparation for supporting narrows types also in generated Java
code according to the shallow embedding approach, when the :GUARDS input is T.