[Java] Hava ATJ generate Java int literals and operations.
The code generation functions of ATJ have been extended to produce Java int
literals and Java int operations from certain ACL2 functions that model the
construction of Java int values and the operations on them. For now only the
binary operations +, -, *, /, and % are supported, but others are easy to add.