Restricted build of books/projects/filesystems/ to "everything" target.
Note that "everything" refers to books/GNUmakefile; the corresponding
target in the top-level ACL2 GNUmakefile is "regression-everything".
This was discussed on slack, and in email with Mihir Mehta.  The file
books/projects/filesystems/lofat.cert has taken more than 2 hours to
certify in one of my recent linux builds (using -j 8 on a 4-core
machine with hyperthreading).

If that book is sped up significantly, then it might be good to remove
books/projects/filesystems/ from the "slow" books that are certified
only with the "everything" target of "make".