Incorporated source code modifications from Sol Swords for speeding up rewriting of certain mv-nth calls.  Eliminated bogus compiler warnings when building ACL2 executable.
This accompanies the preceding commit on "Book changes from Sol Swords
for mv-nth modification.".

Quoting :doc note-8-3:

  A heuristic for ``lazy'' rewriting of calls of mv-nth has been made
  much more efficient in some cases.  Thanks to Sol Swords for
  investigating performance issues leading him to implement such a
  change, and for making modifications to [community-books] so that
  they continue to certify after the change.  Those interested in
  implementation details may start with source function
  simplifiable-mv-nth1.

The change to eliminate bogus compiler warnings is unrelated to the
mv-nth work.  Those compiler warnings started showing up recently in
CCL (but not SBCL, probably because of the heavy-handed way we turn
off warnings in SBCL) due to a change I made to the implementation of
ec-call.