[Java] Add a post-translation phase to ATJ.
This is carried out within Java, after the ACL2-to-Java translation.

Currently post-translation consists of a single step, namely the folding of Java
return statements into branches of preceding if statements, recursively. The
post-translated code is much more idiomatic and uses much fewer local variables;
it may have multiple return statements in different if branches, as opposed to a
single return statement at the end of the method.

All in all, it seems easier to do this folding in a Java-to-Java
post-translation step, rather than as part of the ACL2-to-Java translation: it's
two simpler transformations instead of one more complex transformations.