[Java] Improve code generated by ATJ.
Now we generate synonym methods for imported functions, so that they can be
referenced without qualification as in ACL2.

In ACL2, if F is in package P (i.e. the SYMBOL-PACKAGE-NAME of F is "P"), and
package Q imports P::F, then Q::F and P::F are the same symbol, and in
particular in package Q the unqualified F stands for that symbol, which is much
more readable, especially when Q consists of at least several characters.

With respect to this example, the Java code generated by ATJ includes classes P
and Q for the two packages, and a method f in class P for function F. Before
this change, code in Q had to always use P.f to reference that method. With this
change, we now generate a "synonym" method Q.f that just calls P.f, so that code
in Q can reference that method just as the unqualified Q.

This better mirrors ACL2's notation in Java. It makes the code a lot more
readable. For example, previous occurrences of COMMON_LISP.cons are now just
cons.