[Java] Fix bug in AIJ.
In the initialization of the Java class for ACL2 symbols, the 'COMPLEX symbol
was being created in the "ACL2" package instead of the "COMMON-LISP" package as
it should. For this initial symbol creation, which happens before ACL2 packages
are defined, the exact package (i.e. the one returned by SYMBOL-PACKAGE-NAME)
must be used.