[Java] Optimize code generated by ATJ.
Now we generate Java static final fields for quoted constants that are symbols
and that appear in the bodies of the ACL2 functions translated to Java. This
makes the code faster and more readable. Essentially, it treats these quoted
constants like "literals" that are evaluated once and then just referenced
(recall that these are all objects).

Unlike the static final fields for the other quoted constants, which all appear
in the main Java class, the ones for symbols appear in the nested classes that
correspond to the ACL2 packages. Each such field appears not only in the class
corresponding to the SYMBOL-PACKAGE-NAME of the symbol, but also in the classes
for the packages that import that symbol. Thus, the fields can be referenced in
the Java code according to the same rules as in ACL2 regarding package prefixes
(which in Java are class prefixes).