[Java] Improve ATJ implementation.
Introduce and use an aggregate type for collections of quoted constants, where
CONS pairs are stored into an alist from the CONS pairs to increasing positive
indices, which will be used to generate fields for these constants (in an
upcoming extension).

The aggregate collections are threaded through the code generation functions, as
needed in order to handle the aforementioned indices propertly, but it also
makes the overall code clearer and faster.