Eliminated special variables for "live" stobjs.
I expect that this technical change, which was a bit complicated to
make, will make it relatively easy to implement a new utility,
(swap-stobjs st1 st2).  I'm not pushing this commit separately, as
it's not quite fully tested; I plan to push after adding swap-stobjs.

Also, made trivial efficiency tweak to
output-type-for-declare-form-rec and fixed a minor typo in a warning
message from ev-for-trans-eval.