[Java] Add tail recursion elimination to ATJ.
This is a well-known compiler optimization, which replaces tail recursion with
loops. Now ATJ does that as well.

This is realized as a post-translation, Java-to-Java step, because it seems
simpler to do this than to further complicate the already fairly complicated
ACL2-to-Java translation.

This tail recursion elimination post-translation step relies on the fact that
the ACL2-to-Java translation turns tail-recursive ACL2 functions into
tail-recursive Java methods (not suprisingly so), and also relies on a previous
post-translation steps in which return statements are folded into if statements:
thus, tail recursion elimination looks for recursive calls in return statements.