[Java] Simplify code generated by ATJ.
This concerns the code to initialize (the Java representation of) the ACL2
environment. Now the initialization is done all in the class static initializer
in both the deep and shallow embedding. The method only serves to ensure that
the class is initialized. The initialization field has been removed, as there is
no harm in calling the initialization method multiple times.