[Std/util] Add a new macro STD::DEFFIXER.
This macro automates the definition of fixing functions, and some associated
theorems, according to a common pattern.

A while ago, there was some discussion on the ACL2 Books mailing list about
this. It was suggested to have this macro optionally generate a fixtype, but
after some discussion it was instead agreed to extend FTY::DEFFIXTYPE to
optionally generate the fixer. This macro can be used for that purpose, but it
can also used when a fixer for a predicate is desired, without necessarily using
fixtypes.

This macro can be extended to generate variations of the current form of the
fixer. See :DOC STD::DEFFIXER for details.