[Java] Rename two ATJ type operations.
These are the partial order and the least upper bound on ATJ types. They have
been renamed with an 'a' before 'sub' and 'join' to reflect that they are based
on the ACL2 types denoted by the ATJ types.

There will be a different partial order, and a different semilattice, based on
the Java types denoted by the ATJ types.

These two semilattices play different roles in ATJ.