[Java] Explicate the Java-based semilattice of ATJ types.
Based on the Java types denotes by the ATJ types, define a second partial order
on ATJ types that is an order embedding into the Java subtype relation.  Define
a meet semilattice by adjoining NIL as bottom element. Lift this semilattice to
lists, element-wise. The partial order and meet properties are all proved.

The meet operation on lists will be used to ensure that, when overloaded methods
are generated by (upcoming versions of) ATJ, there is always a clear most
specific method that is resolved at Java compile time.