[Java] Add some operations on ATJ types.
Add a strict version of the Java-base partial order on lists.

Add an operation to find the output type corresponding to the minimum input
types greater than or equal to given types. This is used to see if a function
call can be turned into a call of an overloaded method corresponding to
secondary ATJ function types, and to find the corresponding output type to use
when type-annotating ACL2 terms.