[Java] Extend ATJ type annotation pre-translation step.
Now secondary function types, if present, are taken into account. When we
encounter a function call, we first infer the "natural" types of the actual
arguments, and see if they math (i.e. are less than or equal to) the input types
of some secondary function type. If that's the case, we select the minimum such
secondary function type (which should be guaranteed to exist by the closure
properties enforced by the macro to verify and record secondary function types),
and use the corresponding output type as the inferred type of the function call.
This corresponds to selecting an overloaded method with narrower argument types
(compared to the ones corresponding to the primary function type) at compile
time.