[Std/system] Weaken guard of UBODY+.
So this utility can work better with others like PSEUDO-TERMP.

The check that the symbol argument names a function is now performed at run
time.

The check that the function is in logic mode has been eliminated, so that this
utility can be used for program-mode functions.

The check that the function has a non-NIL UNNORMALIZED BODY property has been
eliminated: we let this utility return NIL in that case, so that the utility can
in fact be used to check whether a function has an UNNORMALIZED-BODY property or
not.