[Java] Extend ATJ to generate Java booleans.
Besides Java ints, now ATJ generates Java code that manipulates Java
booleans. The generation of boolean literals and of all the strict boolean
operations is now supported. These are generated from ACL2 functions that model
the construction of Java boolean values and the application of Java boolean
operations, similarly to what was already done for Java ints. (All of this is
done in the shallow embedding with guards.)

Also added tests to exercise the new capability.