Fixed bug in new stobj hash-table implementation: need to convert rationals to floats.  Fixed an errant "//" in a pathname.
This bug was discovered in GCL; apparently SBCL and CCL are more
forgiving.  The fix is in make-hash-table-with-defaults.  GCL actually
complained about using a rational number as a rehash-threshold, which
seems to me to be legal according to the HyperSpec.  But the HyperSpec
says that rehash-size must be an integer or a float, so now, both are
converted to floats.

Changed an errant "//" to "/" in
books/kestrel/crypto/hmac/hmac-sha-256.lisp (discovered using GCL).