[Java] Fix a bug in ATJ.
Under some circumstances, ATJ generates Java code referencing (the Java
representation of) T or NIL even though these two symbols are not explicitly
used, as quoted constants, in the bodies of the ACL2 functions translated to
Java. Thus, now we initialize the structure that collects quoted constants with
T and NIL, so that final static fields for them will be always generated.