[Java] Extend some AIJ doc and optimize some AIJ code.
Add explicit preconditions for some public AIJ methods to the Javadoc of the
methods. (These are just text; Javadoc does not seem to have constructs for
preconditions as such.)

Add an optimized package-private factory method for ACL2 characters that avoids
the range check on the input Java character, and have AIJ's code call this new
method, instead of the public one, when the input is known to be in range.