[Java] Made several improvements to AIJ.
Introduce and use package-private factory methods for ACL2 values that omit the
defensive checks of the publilc factory methods, since those are always
satisfied internally (under explicit preconditions for the public methods called
by ATJ).

Extend the Javadoc with explicitly stated invariants on non-public fields and on
arguments and results of non-public methoods. These are a further step towards
eventually proving the correctness of the AIJ code.

Made other code and documentation improvements.