[Java] Improve the loops generated by ATJ.
When doing tail recursion elimination, ATJ generates while loops that
destructively update the method parameters just before continuing the loop.
Before this change, all these loops had the form while (true) { ... }, with
termination tests, followed by return's, in the loop body: this is a uniform
tail recursion elimination approach that works for all kinds of tail-recursive
functions, no matter how complex their internal branching (multiple exit tests,
multiple recursive calls, etc.). However, when the body of the while is a
conditional one of whose branches is a return, the if test can be lifted to the
loop (negated if needed), resulting in more idiomatic code.

This is realized as a new Java-to-Java post-translation step that takes place
after the tail recursion elimination step.

It should be possible to lift tests from more complex loop body patterns than
the currently handled ones. This will be done if and when the need arises.