[Java] Add an ATJ test.
This is a tail-recursive Fibonacci function on big integers, which gets
translated to a Java loop. Here the parallel assignment to method parameters
contains circular dependencies, which therefore makes use of temporary local
variables for all the parameters (which actually, with some luck, the JIT might
partially simplify away). This could be a good example for a future, more
nuanced strategy to generate parallel assignments in which temporary local
variables are introduced only for the circular subset of the method parameters.