[Java] Add a formalization of Unicode escapes.
Java's first lexical translation step consists in recognizing Unicode escapes
and turning them into the single Unicode characters that they represent. This
step presents some slight complications, such as the notion of eligible
backslashes, and the fact that under certain conditions the step is required to
fail.

This has been formalized using the relevant rules of the (ABNF) grammar of Java,
as also done in the Java language specification, and by formalizing the
additional English constraints given in the Java language specification (the
grammar rules alone are openly amibiguous).

This first lexical translation step is captured by a function that is
declaratively defined, based on a formalization of the Java language
specification. Future work on this particular part of the Java language
formalization will involve the creation of a verified executable version.