[Java] Improve the Java pretty-printer.
Use postfix as the expected rank of the target of a field access expression. As
noted in the newly added documentation, this does not quite follow the grammar
rules, which require a primary there, but that would prevent expression names
(e.g. expressions of the form var.field, requiring instead (var).field); this
might be an issue with the Java grammar, which should be investigated further.