Replace PSEUDO-FORMP, used in PSEUDO-GOOD-WORLDP.
This is in file [books]/system/pseudo-good-worldp.lisp. The changes to this file
have been written by Matt Kaufmann, but I'm the one creating this commit because
of necessary changes in some Kestrel files noted below. The update consists in
replacing PSEUDO-FORMP, which was used for both event and command forms, and
also allowed atoms, with two new functions PSEUDO-EVENT-FORMP and
PSEUDO-COMMAND-FORMP with more specific checks for events and commands; there
are also the -LISTP versions of both predicates, replacing PSEUDO-FORM-LISTP.

The new functions PSEUDO-EVENT-FORMP and PSEUDO-EVENT-FORM-LISTP are equivalent
to the ones in [books]/kestrel/utilities/system/event-forms.lisp and
[books]/kestrel/utilities/event-form-lists.lisp, which have been tweaked in this
commit to be compatible (i.e. redundant) with the ones in
pseudo-good-worldp.lisp.

This is the 1st step of a 3-step plan in which the new functions will be
factored out of pseudo-good-worldp.lisp, and used instead of the ones in the
Kestrel files, which will be removed/deprecated.

This change was prompted by some discussion with Matt and J about
pseudo-good-worldp.listp, and in particular the PSEUDO-FORMP predicate.