[Java] Remove ATJ-FUNCTION-TYPE->OUTPUT function.
This was "coercing" functions with multiple output types to be treated as
functions with a single :ACONS output type. Now the "coercion" is done by the
(few) callers of this ATJ function, and this ATJ function no longer exists.

Another gradual step towards extending ATJ to support MV values more efficiently
and flexibly.