[Std/system] Add a CHECK-LIST-CALL utility.
This checks whether a term is a translated form of (LIST ...), i.e. is a nest of
CONSes ending in a quoted NIL. If it is, the list of elements (i.e. the CARs) is
returned, in the same order. The term may be just a quoted NIL, in which case it
passes the check and the returned list of elements is empty (NIL).